adds relation prediction and reorganization
